Fix problem on activity screen where incorrect activities were being loaded
The bug was: after logging out and logging back in, the Observable user mode
triggers two calls to the activity feed. This results in either a) double
activity (if the user mode hasn't changed between logins) or b) loading both
Single and Matchmaker activity (if the user mode has changed). This fixes the
symptom, but there still remains the problem that the app is calling the activities
endpoint twice when it should only be calling it once.